pyproject.toml
[tool.poetry]
name = "web8"
version = "0.1.0"
description = ""
authors = ["Tetiana Kobzar <tkobzar.smm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymongo = {extras = ["srv", "tls"], version = "^4.7.2"}
mongoengine = "^0.28.2"
pika = "^1.3.2"
redis = "^5.0.4"
redis-lru = "^0.1.2"
django-contacts = "^0.6.0"
faker = "^25.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
